Re-implement rehype-add-class in TS
